requirements.txt
aiogram~=3.20.0.post0
python-dotenv==0.19.0
aiohttp>=3.10.11
gTTS==2.3.2
googletrans==4.0.0-rc1